Add README for importing CSV into PostgreSQL
This has a much simpler download and ingestion process.  The CSVs are few rather
than many, wrapped in a single zip for download.  An AWK one-liner separates
each CSV into its subtables.  From there, it's easy and fast to import the
tables into PostgreSQL.

This is a dependency of the jobcentre-data repo, which queries the database to
attempt matches by postcode, to find the UPRNs.